requirements.txt
aiohttp
python-dotenv
fastapi
uvicorn
pydantic
requests
pyaudio
keyboard
mutagen
python-multipart
pillow
sqlalchemy
passlib
pymysql
pydantic[email]
bcrypt